extract network collectors
