+++ network/p2p/metrics.go
@@ -1,0 +1,41 @@
+package p2p
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	"log"
+)
+
+var (
+	metricsConnectedPeers = promauto.NewGaugeVec(prometheus.GaugeOpts{
+		Name: "ssv:network:connected_peers",
+		Help: "Count connected peers for a validator",
+	}, []string{"pubKey"})
+	metricsNetMsgsInbound = promauto.NewCounterVec(prometheus.CounterOpts{
+		Name: "ssv:network:net_messages_inbound",
+		Help: "Count incoming network messages",
+	}, []string{"topic"})
+	metricsIBFTMsgsOutbound = promauto.NewCounterVec(prometheus.CounterOpts{
+		Name: "ssv:network:ibft_messages_outbound",
+		Help: "Count IBFT messages outbound",
+	}, []string{"topic"})
+	metricsIBFTDecidedMsgsOutbound = promauto.NewCounterVec(prometheus.CounterOpts{
+		Name: "ssv:network:ibft_decided_messages_outbound",
+		Help: "Count IBFT decided messages outbound",
+	}, []string{"topic"})
+)
+
+func init() {
+	if err := prometheus.Register(metricsConnectedPeers); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+	if err := prometheus.Register(metricsNetMsgsInbound); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+	if err := prometheus.Register(metricsIBFTMsgsOutbound); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+	if err := prometheus.Register(metricsIBFTDecidedMsgsOutbound); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+}
