+++ network/p2p/p2p.go
@@ -5,9 +5,6 @@ import (
 	"crypto/ecdsa"
 	"encoding/json"
 	"fmt"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
-	"log"
 	"net"
 	"sync"
 	"time"
@@ -41,26 +38,6 @@ const (
 
 	syncStreamProtocol = "/sync/0.0.1"
 )
-
-var (
-	metricsConnectedPeers = promauto.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "ssv:network:connected_peers",
-		Help: "Count connected peers for a validator",
-	}, []string{"pubKey"})
-	metricsNetMsgsInbound = promauto.NewCounterVec(prometheus.CounterOpts{
-		Name: "ssv:network:net_messages_inbound",
-		Help: "Count incoming network messages",
-	}, []string{"topic"})
-)
-
-func init() {
-	if err := prometheus.Register(metricsConnectedPeers); err != nil {
-		log.Println("could not register prometheus collector")
-	}
-	if err := prometheus.Register(metricsNetMsgsInbound); err != nil {
-		log.Println("could not register prometheus collector")
-	}
-}
 
 type listener struct {
 	msgCh     chan *proto.SignedMessage
